# src/main/resources/swagger.yml
swagger: "2.0"
info:
  description: "API for icg test"
  version: "1.0.0"
  title: "ICG Test"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "enyert.vinas@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/api/v1/data"
schemes:
  - "https"
paths:
  /creacliente:
    post:
      summary: "Crea un cliente nuevo"
      description: "Crea un cliente nuevo"
      operationId: "creacliente"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Cliente a ser creado"
          required: true
          schema:
            $ref: "#/definitions/Cliente"
      responses:
        405:
          description: "Invalid input"
  /listclientes:
    get:
      summary: "Lista los clientes"
      description: "Lista los clientes"
      operationId: "fetchAllClients"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ClienteResponse"
        400:
          description: "Invalid status value"
  /kpideclientes:
    get:
      summary: "Obtiene kpi de la informacion suministrada de los clientes"
      description: "Obtiene kpi de la informacion suministrada de los clientes"
      operationId: "kpiDeClientes"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/KPIClientes"
        400:
          description: "Invalid tag value"
definitions:
  Cliente:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        format: "string"
      lastName:
        type: "string"
        format: "string"
      age:
        type: "integer"
        format: "int64"
      birthDate:
        type: "string"
        format: "date"
  ClienteResponse:
    type: "object"
    properties:
      cliente:
        $ref: "#/definitions/Cliente"
      approximatedDefuntionDate:
        type: "string"
        format: "date"
  KPIClientes:
    type: "object"
    properties:
      clientsAgeMean:
        type: "number"
        format: "float"
      clientsAgeStandardDeviation:
        type: "number"
        format: "float"